feat: implement local PII redaction and secret scanning with external tests
