+++ Cargo.toml
@@ -19,11 +19,16 @@ chrono = "0.4"
 kube = { version = "3.1.0", features = ["runtime", "client", "derive"], optional = true }
 k8s-openapi = { version = "0.27.1", features = ["v1_31"], optional = true }
 bollard = { version = "0.18.1", optional = true }
+tract-onnx = { version = "0.21.10", optional = true }
+tokenizers = { version = "0.21.0", optional = true }
+ndarray = { version = "0.16.1", optional = true }
+regex = "1.10"
 
 [features]
-default = ["k8s", "docker"]
+default = ["k8s", "docker", "redaction"]
 k8s = ["dep:kube", "dep:k8s-openapi"]
 docker = ["dep:bollard"]
+redaction = ["dep:tract-onnx", "dep:tokenizers", "dep:ndarray"]
 
 [dev-dependencies]
 mockito = "1.4"
