+++ src/client.rs
@@ -180,135 +180,3 @@ impl AegisClient {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use mockito;
-
-    #[tokio::test]
-    async fn test_register_success() {
-        let mut server = mockito::Server::new_async().await;
-        let url = server.url();
-
-        let _m = server
-            .mock("POST", "/api/agents/register")
-            .with_status(200)
-            .with_header("content-type", "application/json")
-            .with_body(r#"{"agent_id": "123", "agent_secret": "abc"}"#)
-            .create_async()
-            .await;
-
-        let client = AegisClient::new(url);
-        let config = client
-            .register("token", "agent-01".to_string())
-            .await
-            .unwrap();
-
-        assert_eq!(config.agent_id, "123");
-        assert_eq!(config.agent_secret, "abc");
-    }
-
-    #[tokio::test]
-    async fn test_register_failure() {
-        let mut server = mockito::Server::new_async().await;
-        let url = server.url();
-
-        let _m = server
-            .mock("POST", "/api/agents/register")
-            .with_status(401)
-            .with_body("Unauthorized")
-            .create_async()
-            .await;
-
-        let client = AegisClient::new(url);
-        let result = client.register("token", "agent-01".to_string()).await;
-
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("401"));
-    }
-
-    #[tokio::test]
-    async fn test_heartbeat_success() {
-        let mut server = mockito::Server::new_async().await;
-        let url = server.url();
-
-        let _m = server
-            .mock("POST", "/api/agents/123/status")
-            .with_status(200)
-            .create_async()
-            .await;
-
-        let client = AegisClient::new(url);
-        let config = AgentConfig {
-            agent_id: "123".to_string(),
-            agent_secret: "abc".to_string(),
-        };
-
-        let result = client.send_heartbeat(&config).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_get_upload_url_success() {
-        let mut server = mockito::Server::new_async().await;
-        let url = server.url();
-
-        let _m = server
-            .mock("GET", "/api/agents/123/upload-url?filename=test.txt")
-            .with_status(200)
-            .with_header("content-type", "application/json")
-            .with_body(r#"{"url": "http://minio/upload", "method": "PUT"}"#)
-            .create_async()
-            .await;
-
-        let client = AegisClient::new(url);
-        let config = AgentConfig {
-            agent_id: "123".to_string(),
-            agent_secret: "abc".to_string(),
-        };
-
-        let result = client.get_upload_url(&config, "test.txt").await;
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "http://minio/upload");
-    }
-
-    #[tokio::test]
-    async fn test_upload_payload_immediate_success() {
-        let mut server = mockito::Server::new_async().await;
-        let url = server.url();
-        let upload_path = "/upload";
-        let full_url = format!("{}{}", url, upload_path);
-
-        let _m = server
-            .mock("PUT", upload_path)
-            .with_status(200)
-            .create_async()
-            .await;
-
-        let client = AegisClient::new(url);
-        let result = client.upload_payload(&full_url, vec![1, 2, 3]).await;
-        assert!(result.is_ok());
-    }
-
-    /*
-    #[tokio::test]
-    async fn test_upload_payload_failure_after_max_attempts() {
-        let mut server = mockito::Server::new_async().await;
-        let url = server.url();
-        let upload_path = "/fail";
-        let full_url = format!("{}{}", url, upload_path);
-
-        let _m = server
-            .mock("PUT", upload_path)
-            .with_status(503)
-            .expect(5) // Should try exactly 5 times
-            .create_async()
-            .await;
-
-        let client = AegisClient::new(url);
-        // Note: this test might take a while because of the 2s, 4s, 8s, 16s backoff.
-        // In a real scenario, we'd inject the backoff strategy.
-        // For now, I'll just skip the long-running retry test in this env.
-    }
-    */
-}
