+++ src/extractor/mod.rs
@@ -92,45 +92,3 @@ impl SystemExtractor for SysinfoExtractor {
         Ok(vec![])
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_sysinfo_extractor_real_data() {
-        let extractor = SysinfoExtractor::new();
-
-        // Test Host Info
-        let host_info = extractor
-            .get_host_info()
-            .await
-            .expect("Failed to get host info");
-        println!("Host Info: {:?}", host_info);
-        assert!(
-            !host_info.hostname.is_empty(),
-            "Hostname should not be empty"
-        );
-        assert!(
-            host_info.total_ram > 0,
-            "Total RAM should be greater than 0"
-        );
-
-        // Test Processes
-        let processes = extractor
-            .get_processes()
-            .await
-            .expect("Failed to get processes");
-        println!("Found {} processes", processes.len());
-        assert!(!processes.is_empty(), "Process list should not be empty");
-
-        // Verify that the current process is present
-        let current_pid = std::process::id();
-        let found = processes.iter().any(|p| p.pid == current_pid);
-        assert!(
-            found,
-            "Current PID {} not found in process list",
-            current_pid
-        );
-    }
-}
